chore(queue): tidy up duplication in queue tests
